% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onze_intercepts_full}
\alias{onze_intercepts_full}
\title{Speaker random intercepts for 418 ONZE speakers}
\format{
A data frame with 481 rows and 21 variables: \describe{
\item{speaker}{Anonymised speaker code.}
\item{F1_DRESS}{Speaker intercept from GAMM model of DRESS F1.}
\item{F2_DRESS}{Speaker intercept from GAMM model of DRESS F2.}
\item{F1_FLEECE}{Speaker intercept from GAMM model of FLEECE F1.}
\item{F2_FLEECE}{Speaker intercept from GAMM model of FLEECE F2.}
\item{F1_GOOSE}{Speaker intercept from GAMM model of GOOSE F1.}
\item{F2_GOOSE}{Speaker intercept from GAMM model of GOOSE F2.}
\item{F1_KIT}{Speaker intercept from GAMM model of KIT F1.}
\item{F2_KIT}{Speaker intercept from GAMM model of KIT F2.}
\item{F1_LOT}{Speaker intercept from GAMM model of LOT F1.}
\item{F2_LOT}{Speaker intercept from GAMM model of LOT F2.}
\item{F1_NURSE}{Speaker intercept from GAMM model of NURSE F1.}
\item{F2_NURSE}{Speaker intercept from GAMM model of NURSE F2.}
\item{F1_START}{Speaker intercept from GAMM model of START F1.}
\item{F2_START}{Speaker intercept from GAMM model of START F2.}
\item{F1_STRUT}{Speaker intercept from GAMM model of STRUT F1.}
\item{F2_STRUT}{Speaker intercept from GAMM model of STRUT F2.}
\item{F1_THOUGHT}{Speaker intercept from GAMM model of THOUGHT F1.}
\item{F2_THOUGHT}{Speaker intercept from GAMM model of THOUGHT F2.}
\item{F1_TRAP}{Speaker intercept from GAMM model of TRAP F1.}
\item{F2_TRAP}{Speaker intercept from GAMM model of TRAP F2.}
}
}
\source{
\url{https://osf.io/q4j29/}
}
\usage{
onze_intercepts_full
}
\description{
A dataset containing the speaker intercepts extracted from GAMM models fit in
Brand et al. (2021).
}
\references{
Brand, James, Jen Hay, Lynn Clark, Kevin Watson & Márton Sóskuthy (2021):
Systematic co-variation of monophthongs across speakers of New Zealand
English. Journal of Phonetics. Elsevier. 88. 101096.
doi:10.1016/j.wocn.2021.101096
}
\keyword{datasets}
