% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_test_plot.R
\name{plot_mds_test}
\alias{plot_mds_test}
\title{Plot \code{\link[=mds_test]{mds_test()}} results}
\usage{
plot_mds_test(mds_test)
}
\arguments{
\item{mds_test}{Object of class \code{mds_test_results} (generated by \code{\link[=mds_test]{mds_test()}}).}
}
\value{
\code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}  Plot output from \code{\link[=mds_test]{mds_test()}}.
}
\examples{
mds_result <- mds_test(
    sim_matrix,
    n_boots = 10,
    n_perms = 10,
    test_dimensions = 3,
    mds_type = 'interval'
 )
 plot_mds_test(mds_result)
}
