% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_matrix}
\alias{sim_matrix}
\title{Similarity matrix from online perception test.}
\format{
A 38x38 matrix
}
\usage{
sim_matrix
}
\description{
Mean similarity ratings for 38 QuakeBox speakers from an online pairwise
similarity task. Random noise added.
}
\keyword{datasets}
