% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_test.R
\name{summary.correlation_test}
\alias{summary.correlation_test}
\title{Summary function for correlation test object. Set alpha to change
significance level.}
\usage{
\method{summary}{correlation_test}(object, alpha = 0.05, n_cors = 5, ...)
}
\arguments{
\item{object}{object of class \verb{correlation test},}

\item{alpha}{significance level for counting correlation as significant.}

\item{n_cors}{number of pairwise correlations to list.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
a \code{glue} object.
}
\description{
Set alpha to change significance level and n_cors to change number of pairwise
correlations given.
}
